requirements.txt
streamlit==1.38.0
Pillow==11.0.0
